requirements.txt
astroid==2.11.2
certifi==2021.10.8
charset-normalizer==2.0.12
defusedxml==0.7.1
dill==0.3.4
idna==3.3
ifaddr==0.1.7
isort==5.10.1
jedi==0.18.1
lazy-object-proxy==1.7.1
librespot==0.0.1
loguru==0.6.0
mccabe==0.7.0
music-tag==0.4.3
mutagen==1.45.1
parso==0.8.3
Pillow==9.1.0
platformdirs==2.5.1
protobuf==3.20.0
pycryptodomex==3.14.1
pydub==0.25.1
pylint==2.13.5
PyOgg==0.6.14a1
python-dotenv==0.20.0
requests==2.27.1
tomli==2.0.1
tqdm==4.64.0
typing_extensions==4.1.1
urllib3==1.26.9
websocket-client==1.3.2
wrapt==1.14.0
zeroconf==0.38.4
